Added process.py; alterations to error checking in separate processes.
process.py now uses a data attribute exitcode rather than a method;
this is in order to emulate the multiprocessing module. I've added
code to classes Process and SGEprocess which is supposed to make sure
that exitcode is None until the process finishes and set appropriately
when thereafter. In the case of SGEprocess, it will only be set
correctly after a call to is_alive returns False: i.e. you must call
is_alive first.