New function pmap: parallel (shell command) map.
Analogous to python standard library map, which applies a python
function to each element of a sequence and returning list of
results. In pmap the python function constructs a shell command for
each element of 'sequence', and a list of files containing the results
of those shell commands is returned. Using classes from process.py,
the shell commands are executed in parallel in a suitable environment:
SGE cluster if settings.sge is True, otherwise separate threads using
the multiprocessing module. I intend to move this function into
process.py, once it has been made fully independent of shellfish.py.