Use -lgfortran in makefile
I think the makefile is pretty much OK now. It was wrong to be using
gfortran instead of gcc; I think what is wanted is -lgfortran with
gcc. I've added -lgfortran to LDFLAGS; that's not ideal because it
means the simplest binaries are linked against lgfortran. So that
could be fixed at some point.